feat(KubeVirt): configure External evictionStrategy on VMs
KubeVirt VMs used for HyperShift must be configured with
the External evictionStrategy.
This setting is essential to ensure that, during a node
drain on the underlying infrastructure cluster
(e.g. during a cluster upgrade),
the eviction request for the KubeVirt VM
is visible to the Cluster API Provider
for KubeVirt (CAPK) controller.

Only when evictionStrategy=External, CAPK can
detect the eviction and initiate a coordinated
drain of the Kubernetes node inside the guest cluster
before the VM is terminated.

If the VM supports live migration, the CAPK controller
will skip draining the guest node, allowing the VM
to be moved without disruption.

For reference, see:
https://github.com/kubernetes-sigs/cluster-api-provider-kubevirt/blob/
2fcbdfd94d2acd3c598cb70c97296377b6a18223/pkg/kubevirt/machine.go#L122-L130

Signed-off-by: Simone Tiraboschi stirabos@redhat.com